Update README.md
This new line(7) is being made to learn about using git. This is all entered in my own repository from github, not WorkingCopy. So far the line is committed directly to my Read-me-edits2 branch.(I was given the option at this point to start a new branch and make a pull request. Now Read-me-edits2 is ahead of main by one commit. Now that I have changes in a branch off of main, I can open a pull request.